Moved graph related stuff to util/graph Moved GeoDataFrameWrapper base class to util/coordinates
